backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.20.0",
    "fastapi>=0.124.4",
    "openai-agents>=0.6.3",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "trafilatura>=2.0.0",
    "uvicorn>=0.38.0",
]
